Algorithms

This section contains the main algorithms of the package.

OAVI

AVI.evaluate_oavi — Method

Applies the OAVI feature transformation to X_test.

AVI.fit_oavi — Method

Creates OAVI feature transformation fitted to X_train

Arguments

'X_train::Union{Matrix{Float64}, Vector{Vector{Float64}}}': training data

'max_degree::Int64': max degree of polynomials computed (default 10)

'psi::Float64': vanishing extent (default 0.1)

'epsilon::Float64': accuracy for convex optimizer (default 1.0e-7)

'tau::Union{Float64, Int64}': upper bound on norm of coefficient vector

'lambda::Float64': regularization parameter

'oracle::Union{String, <:Function}': string denoting which predefined constructor to use OR constructor function. (external constructor function MUST have 'data' and 'labels' as varargs)

'oracle_kwargs::Vector': Array containing keyword arguments for external constructor functions

Returns

'Xtraintransformed::Vector{Vector{Float64}}': transformed X_train

'sets::SetsOandG': instance of 'SetsOandG' keeping track of important sets

VCA

AVI.SetsVCA — Type

Manages the sets V, C and F for VCA.

AVI.F_to_matrix — Method

Transforms SetsVCA.Fs into single matrix.

AVI.V_to_matrix — Method

Transforms SetsVCA.Vs into single matrix.

AVI.apply_V_transformation — Method

Applies transformation corresponding to V polynomials to X_test.

AVI.construct_SetsVCA — Method

Given the data X, this function constructs the intial state of 'SetsVCA'.

AVI.construct_border_vca — Method

constructs the border for current state of the algorithm.

AVI.evaluate_transformation — Method

Evaluates the transformation corresponding to the polynomials in V w.r.t. the functions in F and C

Arguments

'sets::SetsVCA': instance of SetsVCA.

Returns

'totalnumberofzeros::Int64': Sum of all zero entries in coefficientvectors in Vcoefficientvectors.

'totalnumberofentries::Int64': Total number of entries in coefficientvectors in Vcoefficientvectors.

'avgsparsity::Float64': The average sparsity of coefficientvectors in Vcoefficientvectors.

'numberofpolynomials::Int64': Number of polynomials in Vcoefficientvectors.

'numberofterms::Int64': Number of non-vanishing terms.

'degree::Float64': Average degree of polynomials in V.

AVI.evaluate_vca — Method

Evaluates transformation corresponding to the polynomials in V.

AVI.find_range_null_vca — Method

Performs FindRangeNull (using SVD) for VCA. Reference: https://proceedings.mlr.press/v28/livni13.html

Arguments

'F::Union{Matrix{Float64}, Matrix{Int64}}': TBD

'C::Union{Matrix{Float64}, Matrix{Int64}}': TBD

'psi::Float64': vanishing parameter

Returns

'Vcoefficientvectors::Union{Matrix{Float64}, Matrix{Int64}}': Coefficient vectors of polynomials we append to V.

'Vevaluationvectors::Union{Matrix{Float64}, Matrix{Int64}}': Evaluation vectors of polynomials we append to V.

'Fcoefficientvectors::Union{Matrix{Float64}, Matrix{Int64}}': Coefficient vectors of polynomials we append to F.

'Fevaluationvectors::Union{Matrix{Float64}, Matrix{Int64}}': Evaluation vectors of polynomials we append to F.

AVI.fit_vca — Method

This function creates and applies a VCA transformation fitted to X.

Arguments

'X::Matrix{Float64}': data, stored row-wise

'psi::Float64': vanishing parameter

'max_degree::Int64': maximum degree to consider

Returns

'Xtraintransformed::Matrix{Float64}': X transformed according to transformation found by VCA

'sets_vca::SetsVCA': instance of SetsVCA containing relevant sets for VCA

AVI.update_C — Method

updates C

AVI.update_F — Method

updates F sets

AVI.update_V — Method

updates V sets.

Index

AVI.SetsVCA

AVI.F_to_matrix

AVI.V_to_matrix

AVI.apply_V_transformation

AVI.construct_SetsVCA

AVI.construct_border_vca

AVI.evaluate_oavi

AVI.evaluate_transformation

AVI.evaluate_vca

AVI.find_range_null_vca

AVI.fit_oavi

AVI.fit_vca

AVI.update_C

AVI.update_F

AVI.update_V