Terms and polynomials

This section contains information about the handling of terms and polynomials constructed and used during OAVI and VCA. Many of the functions work relatively similar, simply adjusted to the respective structure of storing the information.

Sets $\mathcal{O}$ and $\mathcal{G}$

AVI.SetsOandG — Type

Creates and keeps track of sets O and G for OAVI.

AVI.apply_G_transformation — Method

applies the transformation corresponding to G to X_test

AVI.construct_SetsOandG — Method

initializes SetsOandG instance w.r.t. X_train

AVI.evaluate_transformation — Method

Evaluates the transformation corresponding to the polynomials in Gcoefficientvectors.

Arguments

'sets::SetsOandG': instance of SetsOandG, containing information on transformation

Returns

'totalnumberofzeros::Int64': Sum of all zero entries in coefficientvectors in Gcoefficientvectors.

'totalnumberofentries::Int64': Total number of entries in coefficientvectors in Gcoefficientvectors.

'avgsparsity::Float64': The average sparsity of coefficientvectors in Gcoefficientvectors.

'numberofpolynomials::Int64': Number of polynomials in G.

'numberofterms::Int64': Number of terms in O.

'degree::Float64': Average degree of polynomials in G.

AVI.update_G — Function

updates G sets

AVI.update_O — Method

updates O sets

AVI.update_border — Method

updates border sets

AVI.update_leading_terms — Function

updates leading terms

AVI.update_permutations — Method

updates permutations

Sets $\mathcal{F}$ and $\mathcal{V}$

AVI.SetsVCA — Type

Manages the sets V, C and F for VCA.

AVI.F_to_matrix — Method

Transforms SetsVCA.Fs into single matrix.

AVI.V_to_matrix — Method

Transforms SetsVCA.Vs into single matrix.

AVI.apply_V_transformation — Method

Applies transformation corresponding to V polynomials to X_test.

AVI.construct_SetsVCA — Method

Given the data X, this function constructs the intial state of 'SetsVCA'.

AVI.construct_border_vca — Method

constructs the border for current state of the algorithm.

AVI.evaluate_transformation — Method

Evaluates the transformation corresponding to the polynomials in V w.r.t. the functions in F and C

Arguments

'sets::SetsVCA': instance of SetsVCA.

Returns

'totalnumberofzeros::Int64': Sum of all zero entries in coefficientvectors in Vcoefficientvectors.

'totalnumberofentries::Int64': Total number of entries in coefficientvectors in Vcoefficientvectors.

'avgsparsity::Float64': The average sparsity of coefficientvectors in Vcoefficientvectors.

'numberofpolynomials::Int64': Number of polynomials in Vcoefficientvectors.

'numberofterms::Int64': Number of non-vanishing terms.

'degree::Float64': Average degree of polynomials in V.

AVI.update_C — Method

updates C

AVI.update_F — Method

updates F sets

AVI.update_V — Method

updates V sets.

Index

AVI.SetsOandG

AVI.SetsVCA

AVI.F_to_matrix

AVI.V_to_matrix

AVI.apply_G_transformation

AVI.apply_V_transformation

AVI.construct_SetsOandG

AVI.construct_SetsVCA

AVI.construct_border_vca

AVI.evaluate_transformation

AVI.evaluate_transformation

AVI.update_C

AVI.update_F

AVI.update_G

AVI.update_O

AVI.update_V

AVI.update_border

AVI.update_leading_terms

AVI.update_permutations