Terms and polynomials

This section contains information about the handling of terms and polynomials constructed and used during OAVI and VCA. Many of the functions work relatively similar, simply adjusted to the respective structure of storing the information.

Sets $\mathcal{O}$ and $\mathcal{G}$

ApproximateVanishingIdeals.SetsOandG — Type

Creates and keeps track of sets O and G for OAVI.

ApproximateVanishingIdeals.apply_G_transformation — Method

applies the transformation corresponding to G to X_test

ApproximateVanishingIdeals.construct_SetsOandG — Method

initializes SetsOandG instance w.r.t. X_train

ApproximateVanishingIdeals.evaluate_transformation_oavi — Method

Evaluates the transformation corresponding to the polynomials in Gcoefficientvectors.

Arguments

'sets::SetsOandG': instance of SetsOandG, containing transformation

Returns

'totalnumberofzeros::Int64': Sum of all zero entries in coefficientvectors in Gcoefficientvectors.

'totalnumberofentries::Int64': Total number of entries in coefficientvectors in Gcoefficientvectors.

'avgsparsity::Float64': The average sparsity of coefficientvectors in Gcoefficientvectors.

'numberofpolynomials::Int64': Number of polynomials in G.

'numberofterms::Int64': Number of terms in O.

'degree::Float64': Average degree of polynomials in G.

ApproximateVanishingIdeals.update_G — Function

updates G sets

ApproximateVanishingIdeals.update_O — Method

updates O sets

ApproximateVanishingIdeals.update_border — Method

updates border sets

ApproximateVanishingIdeals.update_leading_terms — Function

updates leading terms

ApproximateVanishingIdeals.update_permutations — Method

updates permutations

Sets $\mathcal{F}$ and $\mathcal{V}$

ApproximateVanishingIdeals.SetsVCA — Type

Manages the sets V, C and F for VCA.

ApproximateVanishingIdeals.F_to_matrix — Method

Transforms SetsVCA.Fs into single matrix.

ApproximateVanishingIdeals.V_to_matrix — Method

Transforms SetsVCA.Vs into single matrix.

ApproximateVanishingIdeals.apply_V_transformation — Method

Applies transformation corresponding to V polynomials to X_test.

ApproximateVanishingIdeals.construct_SetsVCA — Method

Given the data X, this function constructs the intial state of 'SetsVCA'.

ApproximateVanishingIdeals.construct_border_vca — Method

constructs the border for current state of the algorithm.

ApproximateVanishingIdeals.evaluate_transformation_vca — Method

Evaluates the transformation corresponding to the polynomials in V w.r.t. the functions in F and C

Arguments

'sets::SetsVCA': instance of SetsVCA.

Returns

'totalnumberofzeros::Int64': Sum of all zero entries in coefficientvectors in Vcoefficientvectors.

'totalnumberofentries::Int64': Total number of entries in coefficientvectors in Vcoefficientvectors.

'avgsparsity::Float64': The average sparsity of coefficientvectors in Vcoefficientvectors.

'numberofpolynomials::Int64': Number of polynomials in Vcoefficientvectors.

'numberofterms::Int64': Number of non-vanishing terms.

'degree::Float64': Average degree of polynomials in V.

ApproximateVanishingIdeals.update_C — Method

updates C

ApproximateVanishingIdeals.update_F — Method

updates F sets

ApproximateVanishingIdeals.update_V — Method

updates V sets.

Index

ApproximateVanishingIdeals.SetsOandG

ApproximateVanishingIdeals.SetsVCA

ApproximateVanishingIdeals.F_to_matrix

ApproximateVanishingIdeals.V_to_matrix

ApproximateVanishingIdeals.apply_G_transformation

ApproximateVanishingIdeals.apply_V_transformation

ApproximateVanishingIdeals.construct_SetsOandG

ApproximateVanishingIdeals.construct_SetsVCA

ApproximateVanishingIdeals.construct_border_vca

ApproximateVanishingIdeals.evaluate_transformation_oavi

ApproximateVanishingIdeals.evaluate_transformation_vca

ApproximateVanishingIdeals.update_C

ApproximateVanishingIdeals.update_F

ApproximateVanishingIdeals.update_G

ApproximateVanishingIdeals.update_O

ApproximateVanishingIdeals.update_V

ApproximateVanishingIdeals.update_border

ApproximateVanishingIdeals.update_leading_terms

ApproximateVanishingIdeals.update_permutations