Auxiliary functions

This section contains information about auxiliary functions used throughout the algorithm.

AVI specific functions

These functions are used primarily to aid in the computation of the approximate vanishing ideal.

AVI.l1_projection — Method

Projects x onto the L1 Ball with radius 'radius'.

Reference: "Efficient Projections onto the ℓ1-Ball for Learning in High Dimensions", https://stanford.edu/~jduchi/projects/DuchiShSiCh08.pdf

AVI.streaming_matrix_updates — Method

Given A, A.T.A and (A.T.A)^-1 efficiently compute B = [A, a], B.T.B and (B.T.B)^-1. Necessary for fast inverse hessian boosting.

AVI.update_coefficient_vectors — Method

Appends polynomial with coefficient vector based on coefficientvector and term to Gcoefficient_vectors.

Other auxiliary functions

AVI.compute_degree — Method

Computes sum (degree) of columns (terms) in A and returns output as (1 x size(A, 2)) matrix.

AVI.deg_lex_sort — Function

sorts matrix1 degree-lexicographically and matrix2 accordingly

AVI.find_first_non_zero_entries — Method

Finds first non-zero entry in each column and returns a list of the indices. In case of a zero-column returns first index.

AVI.find_last_non_zero_entries — Method

Finds last non-zero entry in each column and returns a list of the indices. In case of a zero-column returns last index.

AVI.get_unique_columns — Function

finds unique columns in matrix x1 and returns only unique elements in x1 as well as corresponding columns in x2.

AVI.orthogonal_projection — Method

Obtains orthogonal projections of vector projected onto vectors.

Printing polynomials

AVI.convert_term_to_latex — Method

Converts a term given as a vector denoting the exponents into LaTeX string

AVI.print_polynomials — Method

Prints the polynomials obtained through OAVI as a LaTeX string.

'digits' can be used to determine how many decimal places you want to round to. Terms with rounded coefficient values 0.0 are omitted and coefficients with value 1.0 are omitted.

Index

AVI.compute_degree

AVI.convert_term_to_latex

AVI.deg_lex_sort

AVI.find_first_non_zero_entries

AVI.find_last_non_zero_entries

AVI.get_unique_columns

AVI.l1_projection

AVI.orthogonal_projection

AVI.print_polynomials

AVI.streaming_matrix_updates

AVI.update_coefficient_vectors